Revert "Changes to improve testing (#9608)" (#9937)
This reverts commit d1a0ccda46d805dead271c7f9c2c41917f51f40a.
