+++ LibreNMS/DB/Eloquent.php
@@ -25,24 +25,21 @@
 
 namespace LibreNMS\DB;
 
-use Dotenv\Dotenv;
 use Illuminate\Database\Capsule\Manager as Capsule;
 use Illuminate\Database\Events\StatementPrepared;
 use Illuminate\Events\Dispatcher;
-use LibreNMS\Util\Laravel;
 
 class Eloquent
 {
     /** @var Capsule static reference to capsule */
     private static $capsule;
+    private static $legacy_listener_installed = false;
 
     public static function boot($options = [])
     {
         // boot Eloquent outside of Laravel
-        if (!Laravel::isBooted() && is_null(self::$capsule)) {
+        if (!defined('LARAVEL_START') && class_exists(Capsule::class) && is_null(self::$capsule)) {
             $install_dir = realpath(__DIR__ . '/../../');
-
-            (new Dotenv($install_dir))->load();
 
             $db_config = include($install_dir . '/config/database.php');
             $settings = $db_config['connections'][$db_config['default']];
@@ -78,7 +75,7 @@ class Eloquent
 
     public static function initLegacyListeners()
     {
-        if (self::isConnected()) {
+        if (self::isConnected() && !self::$legacy_listener_installed) {
             // set FETCH_ASSOC for queries that required by setting the global variable $PDO_FETCH_ASSOC (for dbFacile)
             self::DB()->getEventDispatcher()->listen(StatementPrepared::class, function ($event) {
                 global $PDO_FETCH_ASSOC;
@@ -86,6 +83,7 @@ class Eloquent
                     $event->statement->setFetchMode(\PDO::FETCH_ASSOC);
                 }
             });
+            self::$legacy_listener_installed = true;
         }
     }
 
@@ -109,7 +107,8 @@ class Eloquent
         try {
             $conn = self::DB();
             if ($conn) {
-                return !is_null($conn->getPdo());
+                $conn->getPdo();
+                return true;
             }
         } catch (\PDOException $e) {
             return false;
@@ -126,7 +125,7 @@ class Eloquent
     public static function DB()
     {
         // check if Laravel is booted
-        if (Laravel::isBooted()) {
+        if (defined('LARAVEL_START') && class_exists('DB')) {
             return \DB::connection();
         }
