+++ LibreNMS/Snmptrap/Handlers/UpsmgUtilityRestored.php
@@ -47,6 +47,7 @@ class UpsmgUtilityRestored implements SnmptrapHandler
         }
         $sensor->sensor_current = 2;
         $sensor->save();
-        Log::event("UPS power restored, state sensor " . $sensor->sensor_descr . " has changed to ".$sensor->sensor_current . ".", $device->device_id, "Power", 1);
+        $device_array = $device->toArray();
+        log_event("UPS power restored, state sensor " . $sensor->sensor_descr . " has changed to ".$sensor->sensor_current . ".", $device_array, "Power", 1);
     }
 }
