+++ LibreNMS/Util/Laravel.php
@@ -36,7 +36,7 @@ class Laravel
     public static function bootCli()
     {
         // make sure Laravel isn't already booted
-        if (self::isBooted()) {
+        if (class_exists('App') && App::isBooted()) {
             return;
         }
 
@@ -45,11 +45,6 @@ class Laravel
         $app = require_once $install_dir . '/bootstrap/app.php';
         $kernel = $app->make(\Illuminate\Contracts\Console\Kernel::class);
         $kernel->bootstrap();
-    }
-
-    public static function isBooted()
-    {
-        return !empty(app()->isAlias('Illuminate\Foundation\Application')) && app()->isBooted();
     }
 
     public static function enableQueryDebug()
@@ -67,7 +62,7 @@ class Laravel
                     return $item;
                 })->toJson();
 
-                if (self::isBooted()) {
+                if (class_exists('Log')) {
                     Log::debug("SQL[%Y{$query->sql} %y$bindings%n {$query->time}ms] \n", ['color' => true]);
                 } else {
                     c_echo("SQL[%Y{$query->sql} %y$bindings%n {$query->time}ms] \n");
@@ -88,14 +83,14 @@ class Laravel
 
     public static function enableCliDebugOutput()
     {
-        if (self::isBooted() && App::runningInConsole()) {
+        if (class_exists('\Log') && App::runningInConsole()) {
             Log::setDefaultDriver('console');
         }
     }
 
     public static function disableCliDebugOutput()
     {
-        if (self::isBooted()) {
+        if (class_exists('Log')) {
             Log::setDefaultDriver('logfile');
         }
     }
