+++ app/Http/Controllers/Ajax/NetCommand.php
@@ -26,7 +26,7 @@
 namespace App\Http\Controllers\Ajax;
 
 use App\Http\Controllers\Controller;
-use \LibreNMS\Config;
+use Config;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\StreamedResponse;
 use Symfony\Component\Process\Process;
