+++ config/database.php
@@ -25,7 +25,7 @@ return [
     |
     */
 
-    'default' => env('DB_CONNECTION', env('DBTEST') ? 'testing' : 'mysql'),
+    'default' => env('DB_CONNECTION', 'mysql'),
 
     /*
     |--------------------------------------------------------------------------
@@ -64,21 +64,6 @@ return [
             'collation' => 'utf8_unicode_ci',
             'prefix' => '',
             'prefix_indexes' => true,
-            'strict' => true,
-            'engine' => null,
-        ],
-
-        'testing' => [
-            'driver' => env('DB_TEST_DRIVER', 'mysql'),
-            'host' => env('DB_TEST_HOST', 'localhost'),
-            'port' => env('DB_TEST_PORT', ''),
-            'database' => env('DB_TEST_DATABASE', 'librenms_phpunit_78hunjuybybh'),
-            'username' => env('DB_TEST_USERNAME', 'root'),
-            'password' => env('DB_TEST_PASSWORD', ''),
-            'unix_socket' => env('DB_TEST_SOCKET', ''),
-            'charset' => 'utf8',
-            'collation' => 'utf8_unicode_ci',
-            'prefix' => '',
             'strict' => true,
             'engine' => null,
         ],
