+++ includes/helpers.php
@@ -23,8 +23,6 @@
  * @author     Tony Murray <murraytony@gmail.com>
  */
 
-use LibreNMS\Util\Laravel;
-
 if (!function_exists('d_echo')) {
     /**
      * Legacy convenience function - please use this instead of 'if ($debug) { echo ...; }'
@@ -37,7 +35,7 @@ if (!function_exists('d_echo')) {
     {
         global $debug;
 
-        if (Laravel::isBooted()) {
+        if (class_exists('\Log')) {
             \Log::debug(is_string($text) ? rtrim($text) : $text);
         } elseif ($debug) {
             print_r($text);
