+++ includes/polling/ipmi.inc.php
@@ -1,6 +1,5 @@
 <?php
 
-use LibreNMS\Config;
 use LibreNMS\RRD\RrdDefinition;
 
 $ipmi_rows = dbFetchRows("SELECT * FROM sensors WHERE device_id = ? AND poller_type='ipmi'", array($device['device_id']));
