+++ phpunit.xml
@@ -39,6 +39,5 @@
         <env name="MAIL_DRIVER" value="array"/>
         <env name="QUEUE_CONNECTION" value="sync"/>
         <env name="SESSION_DRIVER" value="array"/>
-        <env name="DB_CONNECTION" value="testing"/>
     </php>
 </phpunit>
