+++ resources/views/locations.blade.php
@@ -148,7 +148,7 @@
                 locationId = $btn.data('id');
 
                 if (locationMap === null) {
-                    config = {"tile_url": "{{ \LibreNMS\Config::get('leaflet.tile_url', '{s}.tile.openstreetmap.org') }}"};
+                    config = {"tile_url": "{{ Config::get('leaflet.tile_url', '{s}.tile.openstreetmap.org') }}"};
                     locationMap = init_map('location-edit-map', '{{ $maps_engine }}', '{{ $maps_api }}', config);
                     locationMarker = init_map_marker(locationMap, location);
                 }
