+++ resources/views/widgets/graylog.blade.php
@@ -19,11 +19,11 @@
         rowCount: ['{{ $limit }}', 25,50,100,250,-1],
         formatters: {
             "browserTime": function(column, row) {
-                @config('graylog.timezone')
+                @if(Config::get('graylog.timezone'))
                     return row.timestamp;
                 @else
                     return moment.parseZone(row.timestamp).local().format("YYYY-MM-DD HH:MM:SS");
-                @endconfig
+                @endif
             }
         },
         post: function ()
