+++ tests/TestCase.php
@@ -42,12 +42,6 @@ abstract class TestCase extends \PHPUnit\Framework\TestCase
         $this->snmpsim = $snmpsim;
     }
 
-    public function setUp()
-    {
-        parent::setUp();
-        set_debug(false); // prevent warnings from stopping execution for legacy code
-    }
-
     public function dbSetUp()
     {
         if (getenv('DBTEST')) {
