+++ tests/snmpsim/fabos_1.snmprec
@@ -1,2 +1,2 @@
-1.3.6.1.2.1.1.1.0|4|FC5022 16Gb SAN Scalable Switch:IBM Flex System FC5022 24-port 16Gb ESB SAN Scalable Switch
+1.3.6.1.2.1.1.2.0|4|FC5022 16Gb SAN Scalable Switch:IBM Flex System FC5022 24-port 16Gb ESB SAN Scalable Switch
 1.3.6.1.2.1.1.2.0|6|1.3.6.1.4.1.1588.2.1.1.117
